some changes

+++ CampusJungle/Controllers/CreateClassController/CCCreateClassController.m
@@ -227,7 +227,7 @@
 - (void)deleteCellObject:(NSDictionary *)object{
     NSInteger objectIndex = [self.tableDataProvider.arrayOfLessons indexOfObject:object];
     NSIndexPath *indexPath = [NSIndexPath indexPathForRow:objectIndex inSection:0];
-    [(NSMutableArray *)self.tableDataProvider.arrayOfItems removeObjectAtIndex:objectIndex];
+    [(NSMutableArray *) self.tableDataProvider.arrayOfItems removeObjectAtIndex:(NSUInteger) objectIndex];
     [self.mainTable beginUpdates];
     [self.mainTable deleteRowsAtIndexPaths:@[indexPath] withRowAnimation:UITableViewRowAnimationFade];
     [self.mainTable endUpdates];
